pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["db", "module1", "module2", "module3"]

[project]
name = "jobo"
version = "0.1.0"
description = "Job Application Automation System"
requires-python = ">=3.11"
dependencies = [
    "click",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "sqlalchemy",
    "beautifulsoup4",
    "lxml",
    "python-dateutil",
    "python-dotenv",
    "playwright",
    "httpx",
    "pydantic",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
jobo = "main:cli"

[tool.ruff]
line-length = 88
